+++ doc/src/modules/codegen.rst
@@ -615,10 +615,3 @@ Fortran utilities (sympy.codegen.futils)
 
 .. automodule:: sympy.codegen.futils
    :members:
-
-
-Array utilities (sympy.codegen.array_utils)
--------------------------------------------
-
-.. automodule:: sympy.codegen.array_utils
-   :members:
