+++ sympy/codegen/tests/test_array_utils.py
@@ -1,18 +1,18 @@
 from sympy import (
     symbols, IndexedBase, Identity, cos, Inverse, tensorcontraction,
     permutedims, tensorproduct, ZeroMatrix, HadamardProduct, HadamardPower, MatPow, sin)
-from sympy.codegen.array_utils import (
-    CodegenArrayContraction, CodegenArrayTensorProduct, CodegenArrayDiagonal,
-    CodegenArrayPermuteDims, CodegenArrayElementwiseAdd, _codegen_array_parse,
-    parse_indexed_expression, recognize_matrix_expression,
-    parse_matrix_expression, nest_permutation, _remove_trivial_dims, _array_diag2contr_diagmatrix,
-    _support_function_tp1_recognize, array2matrix, ArrayElementwiseApplyFunc)
+from sympy.tensor.array.expressions.conv_matrix_to_array import parse_matrix_expression
+from sympy.tensor.array.expressions.conv_array_to_matrix import _support_function_tp1_recognize, \
+    _array_diag2contr_diagmatrix, recognize_matrix_expression, _remove_trivial_dims, array2matrix
+from sympy.tensor.array.expressions.conv_indexed_to_array import parse_indexed_expression, _convert_indexed_to_array
 from sympy import MatrixSymbol, Sum
 from sympy.combinatorics import Permutation
 from sympy.functions.special.tensor_functions import KroneckerDelta
 from sympy.matrices.expressions.diagonal import DiagMatrix
 from sympy.matrices import Trace, MatMul, Transpose
-from sympy.tensor.array.expressions.array_expressions import ZeroArray, OneArray, ArraySymbol
+from sympy.tensor.array.expressions.array_expressions import ZeroArray, OneArray, ArraySymbol, \
+    CodegenArrayTensorProduct, CodegenArrayElementwiseAdd, CodegenArrayPermuteDims, CodegenArrayDiagonal, \
+    ArrayElementwiseApplyFunc, CodegenArrayContraction, nest_permutation
 from sympy.testing.pytest import raises
 import random
 
@@ -213,43 +213,43 @@ def test_codegen_array_flatten():
 
 def test_codegen_array_parse():
     expr = M[i, j]
-    assert _codegen_array_parse(expr) == (M, (i, j))
+    assert _convert_indexed_to_array(expr) == (M, (i, j))
     expr = M[i, j]*N[k, l]
-    assert _codegen_array_parse(expr) == (CodegenArrayTensorProduct(M, N), (i, j, k, l))
+    assert _convert_indexed_to_array(expr) == (CodegenArrayTensorProduct(M, N), (i, j, k, l))
     expr = M[i, j]*N[j, k]
-    assert _codegen_array_parse(expr) == (CodegenArrayDiagonal(CodegenArrayTensorProduct(M, N), (1, 2)), (i, k, j))
+    assert _convert_indexed_to_array(expr) == (CodegenArrayDiagonal(CodegenArrayTensorProduct(M, N), (1, 2)), (i, k, j))
     expr = Sum(M[i, j]*N[j, k], (j, 0, k-1))
-    assert _codegen_array_parse(expr) == (CodegenArrayContraction(CodegenArrayTensorProduct(M, N), (1, 2)), (i, k))
+    assert _convert_indexed_to_array(expr) == (CodegenArrayContraction(CodegenArrayTensorProduct(M, N), (1, 2)), (i, k))
     expr = M[i, j] + N[i, j]
-    assert _codegen_array_parse(expr) == (CodegenArrayElementwiseAdd(M, N), (i, j))
+    assert _convert_indexed_to_array(expr) == (CodegenArrayElementwiseAdd(M, N), (i, j))
     expr = M[i, j] + N[j, i]
-    assert _codegen_array_parse(expr) == (CodegenArrayElementwiseAdd(M, CodegenArrayPermuteDims(N, Permutation([1,0]))), (i, j))
+    assert _convert_indexed_to_array(expr) == (CodegenArrayElementwiseAdd(M, CodegenArrayPermuteDims(N, Permutation([1, 0]))), (i, j))
     expr = M[i, j] + M[j, i]
-    assert _codegen_array_parse(expr) == (CodegenArrayElementwiseAdd(M, CodegenArrayPermuteDims(M, Permutation([1,0]))), (i, j))
+    assert _convert_indexed_to_array(expr) == (CodegenArrayElementwiseAdd(M, CodegenArrayPermuteDims(M, Permutation([1, 0]))), (i, j))
     expr = (M*N*P)[i, j]
-    assert _codegen_array_parse(expr) == (CodegenArrayContraction(CodegenArrayTensorProduct(M, N, P), (1, 2), (3, 4)), (i, j))
+    assert _convert_indexed_to_array(expr) == (CodegenArrayContraction(CodegenArrayTensorProduct(M, N, P), (1, 2), (3, 4)), (i, j))
     expr = expr.function  # Disregard summation in previous expression
-    ret1, ret2 =  _codegen_array_parse(expr)
+    ret1, ret2 =  _convert_indexed_to_array(expr)
     assert ret1 == CodegenArrayDiagonal(CodegenArrayTensorProduct(M, N, P), (1, 2), (3, 4))
     assert str(ret2) == "(i, j, _i_1, _i_2)"
     expr = KroneckerDelta(i, j)*M[i, k]
-    assert _codegen_array_parse(expr) == (M, ({i, j}, k))
+    assert _convert_indexed_to_array(expr) == (M, ({i, j}, k))
     expr = KroneckerDelta(i, j)*KroneckerDelta(j, k)*M[i, l]
-    assert _codegen_array_parse(expr) == (M, ({i, j, k}, l))
+    assert _convert_indexed_to_array(expr) == (M, ({i, j, k}, l))
     expr = KroneckerDelta(j, k)*(M[i, j]*N[k, l] + N[i, j]*M[k, l])
-    assert _codegen_array_parse(expr) == (CodegenArrayDiagonal(CodegenArrayElementwiseAdd(
+    assert _convert_indexed_to_array(expr) == (CodegenArrayDiagonal(CodegenArrayElementwiseAdd(
             CodegenArrayTensorProduct(M, N),
             CodegenArrayPermuteDims(CodegenArrayTensorProduct(M, N), Permutation(0, 2)(1, 3))
         ), (1, 2)), (i, l, frozenset({j, k})))
     expr = KroneckerDelta(j, m)*KroneckerDelta(m, k)*(M[i, j]*N[k, l] + N[i, j]*M[k, l])
-    assert _codegen_array_parse(expr) == (CodegenArrayDiagonal(CodegenArrayElementwiseAdd(
+    assert _convert_indexed_to_array(expr) == (CodegenArrayDiagonal(CodegenArrayElementwiseAdd(
             CodegenArrayTensorProduct(M, N),
             CodegenArrayPermuteDims(CodegenArrayTensorProduct(M, N), Permutation(0, 2)(1, 3))
         ), (1, 2)), (i, l, frozenset({j, m, k})))
     expr = KroneckerDelta(i, j)*KroneckerDelta(j, k)*KroneckerDelta(k,m)*M[i, 0]*KroneckerDelta(m, n)
-    assert _codegen_array_parse(expr) == (M, ({i,j,k,m,n}, 0))
+    assert _convert_indexed_to_array(expr) == (M, ({i, j, k, m, n}, 0))
     expr = M[i, i]
-    assert _codegen_array_parse(expr) == (CodegenArrayDiagonal(M, (0, 1)), (i,))
+    assert _convert_indexed_to_array(expr) == (CodegenArrayDiagonal(M, (0, 1)), (i,))
 
 
 def test_codegen_array_diagonal():
@@ -266,29 +266,29 @@ def test_codegen_recognize_matrix_expression():
     assert recognize_matrix_expression(expr) == M + Transpose(M)
 
     expr = M[i,j] + N[i,j]
-    p1, p2 = _codegen_array_parse(expr)
+    p1, p2 = _convert_indexed_to_array(expr)
     assert recognize_matrix_expression(p1) == M + N
 
     expr = M[i,j] + N[j,i]
-    p1, p2 = _codegen_array_parse(expr)
+    p1, p2 = _convert_indexed_to_array(expr)
     assert recognize_matrix_expression(p1) == M + N.T
 
     expr = M[i,j]*N[k,l] + N[i,j]*M[k,l]
-    p1, p2 = _codegen_array_parse(expr)
+    p1, p2 = _convert_indexed_to_array(expr)
     assert recognize_matrix_expression(p1) == CodegenArrayElementwiseAdd(
         CodegenArrayTensorProduct(M, N),
         CodegenArrayTensorProduct(N, M))
 
     expr = (M*N*P)[i, j]
-    p1, p2 = _codegen_array_parse(expr)
+    p1, p2 = _convert_indexed_to_array(expr)
     assert recognize_matrix_expression(p1) == M*N*P
 
     expr = Sum(M[i,j]*(N*P)[j,m], (j, 0, k-1))
-    p1, p2 = _codegen_array_parse(expr)
+    p1, p2 = _convert_indexed_to_array(expr)
     assert recognize_matrix_expression(p1) == M*N*P
 
     expr = Sum((P[j, m] + P[m, j])*(M[i,j]*N[m,n] + N[i,j]*M[m,n]), (j, 0, k-1), (m, 0, k-1))
-    p1, p2 = _codegen_array_parse(expr)
+    p1, p2 = _convert_indexed_to_array(expr)
     assert recognize_matrix_expression(p1) == M*P*N + M*P.T*N + N*P*M + N*P.T*M
